feat: 보너스 번호 검사 함수 추가
